backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["J-Yunim <2366374422@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
langchain = "^0.3.13"
qdrant-client = "^1.12.1"
langchain-qdrant = "^0.2.0"
python-decouple = "^3.8"
langchain-community = "^0.3.13"
openai = "^1.58.1"
langchain-openai = "^0.2.14"
langchain-ollama = "^0.2.2"
bs4 = "^0.0.2"
langgraph = "^0.2.60"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
